+++ src/index.ts
@@ -1,2 +1,2 @@
-export * from './cfcore.js'
 export * from './CFCoreClient.js'
+export * from './types.js'